refactor(filter): move classify module to classifier directory
Promote the flat classify.rs file to a classifier/ directory module,
aligning with the directory-module pattern used by request_validate/.
